+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Improved positioning of ornaments with multiple layers
 * Improved articulations in normal and cross-staff contexts
 * Improved spacing of clef changes
+* Improved timestamp events alignments by avoiding extra-space to be added
 * Embedding of Leland font from MuseScore
 * Method Toolkit::GetTimesForElement for accessing score times (@wergo)
 * Option for sometimes using encoded breaks, at configurable threshold (`--breaks smart` and `--breaks-smart-sb`) (@earboxer)